embed horapg helpers into bet
