+++ index.js
@@ -11,9 +11,8 @@ const FormData = require('form-data');
 const chalk = require('chalk');
 const moment = require('moment-timezone');
 const config = require('./dono/config.json');
-const { obterHorarioAtual, buscarHorarios, verificarHorariosEEnviarMensagens } = require('./func/bet.js');
+const { obterHorarioAtual, buscarHorarios, verificarHorariosEEnviarMensagens, storeHorapg, updateLastSent } = require('./func/bet.js');
 const { processTikTokMedia, processKwaiMedia, downloadVideoFromYouTube, downloadFromApi } = require('./func/downloader.js');
-const { storeHorapg, updateLastSent } = require('./func/horapg.js');
 const os = require('os');
 const ping = require('ping');
 const ffmpeg = require('fluent-ffmpeg');
